moving split_spender_key to frost_utils
