+++ ironfish-rust/src/frost_utils/mod.rs
@@ -4,3 +4,4 @@
 
 pub mod round_one;
 pub mod split_secret;
+pub mod split_spender_key;
